### 2026-02-02 7cc9679
docs: fix issues flagged by linter
