+++ src/kooplearn/datasets/_overdamped_langevin_generator.py
@@ -557,18 +557,19 @@ def compute_prinz_potential_eig(gamma, sigma, dt, eval_right_on=None, num_compon
     Returns
     -------
     eigenvalues : ndarray
-        The ``num_components`` dominant Koopman eigenvalues (i.e., :math:`e^{\\lambda_i \\Delta t}`),
-        sorted by magnitude in ascending order. Shape ``(num_components,)``.
+        The ``num_components`` dominant Koopman eigenvalues
+        (i.e., :math:`e^{\\lambda_i \\Delta t}`), sorted by magnitude in ascending order.
+        Shape ``(num_components,)``.
     eigenfunctions : ndarray
-        The leading ``num_components`` right eigenfunctions evaluated at `eval_right_on`` if not ``None``.
-        Shape ``(len(eval_right_on), num_components)``.
+        The leading ``num_components`` right eigenfunctions evaluated at `eval_right_on`` if not
+        ``None``. Shape ``(len(eval_right_on), num_components)``.
 
     Notes
     -----
     The computation uses a Galerkin projection method with a 1D cosine basis
     to discretize the Fokker-Planck operator. The domain for the basis functions
     is set to ``(-3, 3)``.
-    """
+    """  # noqa: E501
     prinz_grad = lambda x: (
         -128 * np.exp(-80 * ((-0.5 + x) ** 2)) * (-0.5 + x)
         - 512 * np.exp(-80 * (x**2)) * x
