### 2026-02-02 67eec94
docs: add kooplearn bibtex citation
