### 2026-02-02 b313961
docs: fix bibliography
