+++ docs/bibliography.bib
@@ -1,20 +1,26 @@
-
-@article{Kostic2022,
-  title={Learning dynamical systems via koopman operator regression in reproducing kernel hilbert spaces},
-  author={Kostic, Vladimir and Novelli, Pietro and Maurer, Andreas and Ciliberto, Carlo and Rosasco, Lorenzo and Pontil, Massimiliano},
-  journal={Advances in Neural Information Processing Systems},
-  volume={35},
-  pages={4017--4031},
-  year={2022}
+@inproceedings{Kostic2022,
+  author = {Kostic, Vladimir R. and Novelli, Pietro and Maurer, Andreas and Ciliberto, Carlo and Rosasco, Lorenzo and Pontil, Massimiliano},
+  booktitle = {Advances in Neural Information Processing Systems},
+  editor = {S. Koyejo and S. Mohamed and A. Agarwal and D. Belgrave and K. Cho and A. Oh},
+  pages = {4017--4031},
+  publisher = {Curran Associates, Inc.},
+  title = {Learning Dynamical Systems via Koopman Operator Regression in Reproducing Kernel Hilbert Spaces},
+  url = {https://proceedings.neurips.cc/paper_files/paper/2022/file/196c4e02b7464c554f0f5646af5d502e-Paper-Conference.pdf},
+  volume = {35},
+  year = {2022}
 }
 
-@article{kostic2024learning,
-  title={Learning the infinitesimal generator of stochastic diffusion processes},
-  author={Kostic, Vladimir and Halconruy, H{\'e}l{\`e}ne and Devergne, Timoth{\'e}e and Lounici, Karim and Pontil, Massimiliano},
-  journal={Advances in Neural Information Processing Systems},
-  volume={37},
-  pages={137806--137846},
-  year={2024}
+@inproceedings{kostic2024learning,
+  author = {Kostic, Vladimir R. and Lounici, Karim and Halconruy, H\'{e}l\`{e}ne and Devergne, Timoth\'{e}e and Pontil, Massimiliano},
+  booktitle = {Advances in Neural Information Processing Systems},
+  doi = {10.52202/079017-4377},
+  editor = {A. Globerson and L. Mackey and D. Belgrave and A. Fan and U. Paquet and J. Tomczak and C. Zhang},
+  pages = {137806--137846},
+  publisher = {Curran Associates, Inc.},
+  title = {Learning the Infinitesimal Generator of Stochastic Diffusion Processes},
+  url = {https://proceedings.neurips.cc/paper_files/paper/2024/file/f930c6e1292a1160189a0734f22b465b-Paper-Conference.pdf},
+  volume = {37},
+  year = {2024}
 }
 
 @book{Applebaum2009,
@@ -53,16 +59,18 @@
   publisher={AIP Publishing}
 }
 
-
 @article{Koopman1931,
-  title={Hamiltonian systems and transformation in Hilbert space},
-  author={Koopman, Bernard O},
-  journal={Proceedings of the National Academy of Sciences},
-  volume={17},
-  number={5},
-  pages={315--318},
-  year={1931},
-}
+  author = {B. O. Koopman },
+  title = {Hamiltonian Systems and Transformation in Hilbert Space},
+  journal = {Proceedings of the National Academy of Sciences},
+  volume = {17},
+  number = {5},
+  pages = {315-318},
+  year = {1931},
+  doi = {10.1073/pnas.17.5.315},
+  URL = {https://www.pnas.org/doi/abs/10.1073/pnas.17.5.315},
+  eprint = {https://www.pnas.org/doi/pdf/10.1073/pnas.17.5.315}
+  }
 
 @article{Prinz2011,
   title = {Markov models of molecular kinetics: Generation and validation},
@@ -78,27 +86,33 @@
   month = may 
[…]
