### 2026-02-02 9c4582d
docs: update JOSS paper
