+++ docs/JOSS/paper.bib
@@ -1,39 +1,59 @@
-@article{kostic2024generator,
-  title={Learning the infinitesimal generator of stochastic diffusion processes},
-  author={Kostic, Vladimir R and Halconruy, H{\'e}l{\`e}ne and Devergne, Timoth{\'e}e and Lounici, Karim and Pontil, Massimiliano},
-  journal={Advances in Neural Information Processing Systems},
-  volume={37},
-  pages={137806--137846},
-  year={2024}
+@inproceedings{kostic2024generator,
+  author = {Kostic, Vladimir R. and Lounici, Karim and Halconruy, H\'{e}l\`{e}ne and Devergne, Timoth\'{e}e and Pontil, Massimiliano},
+  booktitle = {Advances in Neural Information Processing Systems},
+  doi = {10.52202/079017-4377},
+  editor = {A. Globerson and L. Mackey and D. Belgrave and A. Fan and U. Paquet and J. Tomczak and C. Zhang},
+  pages = {137806--137846},
+  publisher = {Curran Associates, Inc.},
+  title = {Learning the Infinitesimal Generator of Stochastic Diffusion Processes},
+  url = {https://proceedings.neurips.cc/paper_files/paper/2024/file/f930c6e1292a1160189a0734f22b465b-Paper-Conference.pdf},
+  volume = {37},
+  year = {2024}
 }
 
-@article{devergne2024biased,
-  title={From biased to unbiased dynamics: An infinitesimal generator approach},
-  author={Devergne, Timoth{\'e}e and Kostic, Vladimir R and Parrinello, Michele and Pontil, Massimiliano},
-  journal={Advances in Neural Information Processing Systems},
-  volume={37},
-  pages={75495--75521},
-  year={2024}
+@inproceedings{devergne2024biased,
+  author = {Devergne, Timoth\'{e}e and Kostic, Vladimir R. and Parrinello, Michele and Pontil, Massimiliano},
+  booktitle = {Advances in Neural Information Processing Systems},
+  doi = {10.52202/079017-2404},
+  editor = {A. Globerson and L. Mackey and D. Belgrave and A. Fan and U. Paquet and J. Tomczak and C. Zhang},
+  pages = {75495--75521},
+  publisher = {Curran Associates, Inc.},
+  title = {From Biased to Unbiased Dynamics: An Infinitesimal Generator Approach},
+  url = {https://proceedings.neurips.cc/paper_files/paper/2024/file/89edef87915d31de3437b6b2ac5f79e7-Paper-Conference.pdf},
+  volume = {37},
+  year = {2024}
 }
 
 @article{devergne2025slow,
-  title={Slow dynamical modes from static averages},
-  author={Devergne, Timoth{\'e}e and Kostic, Vladimir R and Pontil, Massimiliano and Parrinello, Michele},
-  journal={The Journal of Chemical Physics},
-  volume={162},
-  number={12},
-  year={2025},
-  publisher={AIP Publishing}
+  author = {Devergne, Timoth{\'e}e and Kostic, Vladimir R. and Pontil, Massimiliano and Parrinello, Michele},
+  title = {Slow dynamical modes from static averages},
+  journal = {The Journal of Chemical Physics},
+  volume = {162},
+  number = {12},
+  pages = {124108},
+  year = {2025},
+  month = {03},
[…]
