+++ docs/JOSS/paper.md
@@ -32,7 +32,7 @@ authors:
   - name: Karim Lounici
     orcid: 0000-0001-6806-6303
     affiliation: 2
-  - name: Vladimir Kostic
+  - name: Vladimir R. Kostic
     orcid: 0000-0002-2876-8834
     affiliation: "1, 6"
   - name: Massimiliano Pontil
@@ -70,17 +70,19 @@ From fluid flows down to atomistic motions, dynamical systems permeate every sci
 $$(\mathsf{E} f)(x_t) = \int p(dy | x_{t}) f(y) = \mathbb{E}_{y \sim X_{t+1} | X_{t}}[f(y) | x_t].$$
 Notice that $\mathsf{E}$ is an operator because it maps any function $f$ to another function, $x_{t} \mapsto (\mathsf{E} f)(x_t)$, and is *linear* because $\mathsf{E}(f + \alpha g) = \mathsf{E} f + \alpha \mathsf{E} g$. When the dynamics is deterministic, $\mathsf{E}$ is known as the *Koopman operator* [@koopman1931], while in the stochastic case it is known as the *transfer operator* [@applebaum2009]. Arguably, the most important feature of evolution operators is their spectral decomposition [@mezic2005], which can be used to express the dynamics as a linear superposition of *modes*. These ideas lie at the core of the celebrated Time-lagged Independent Component Analysis [@Molgedey1994], and  Dynamical Mode Decomposition (DMD) [@schmid2010; @kutz2016]. 
 
-Evolution operator learning is best understood from the perspective of *latent linear dynamical models*, which is schematically depicted in \autoref{fig:evop_scheme}. In this framework, the dynamical state $x_t$ is first mapped into a latent space defined by a (fixed or learned) representation $\varphi$. Then, a *linear evolution* map $E$ is learned to approximate the dynamics of the latents. The pair $(\varphi, E)$ provides an approximation of $\mathsf{E}$ restricted to the $d$-dimensional subspace spanned by the components of $\varphi$, given the data. `kooplearn` implements state-of-the-art methods to learn $\varphi$, $E$, and the associated spectral decomposition of $\mathsf{E}$. 
+Evolution operator learning is best understood from the perspective of *latent linear dynamical models*, which is schematically depicted in \autoref{fig:evop_scheme}. In this framework, the dynamical state $x_t$ is first mapped into a latent space defined by a (fixed or learned) representation $\varphi$. Then, a *linear evolution* map $E$ is learned to approximate the dynamics of the latents. The pair $(\varphi, E)$ provides an approximation of $\mathsf{E}$ restricted to the $d$-dimensional subspace spanned by the components of $\varphi$, given the data. `kooplearn` implements methods to learn $\varphi$, $E$, and the associated spectral decomposition of $\mathsf{E}$. 
 
-![Sketch of the action of an evolution operator on a protein folding trajectory. The dynamics of the protein is linearized by means of a nonlinear representation $\varphi$ and subsequently evolved by means of the linear map $E$.\label{fig:evop_scheme}](Fig1.png){ width=100% }
+![Sketch of the action of an evolution operator on a protein folding trajectory. The dynamics of the protein is linearized by means of a nonlinear representation $\varphi$ and subsequently evolved by means of the linear map $E$.\label{fig:evop_scheme}](Fig1.png){ width=90% }
+
+# State of the Field
 
[…]
