### 2026-02-02 #9787 @g-turri
Hi @crvernon,

Thank you for pointing out the new requirements. We have amended the paper to include the `Software Design` and the `Research Impact Statement` sections. We also added a  `State of the Field` section.

I would also like to suggest the following potential reviewers:
- dario-coscia
- fAndreuzzi
- garrett-m-smith
- matteosantama
