### 2026-02-06 8f994ce
docs: fixed typo
