### 2026-02-10 #9787 @gmarupilla
@crvernon, I’d like to volunteer to review this paper. I’ve worked with ML/AI and scientific Python tools and I’m happy to evaluate installation, functionality, documentation, and reproducibility. Please let me know if you’d like me to proceed.
