### 2026-04-01 #10342 @editorialbot
[REVIEW]: kooplearn: A Scikit-Learn Compatible Library of Algorithms for Evolution Operator Learning
**Submitting author:** @g-turri (<a href="http://orcid.org/0000-0002-3405-9292">Giacomo Turri</a>)
**Repository:** https://github.com/Machine-Learning-Dynamical-Systems/kooplearn
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.2
