### 2026-04-07 Machine-Learning-Dynamical-Systems/kooplearn#24 @sdahdah
Two example notebooks do not render in GitHub
`ala2_nys_tutorial.ipynb` and `kernel_methods.ipynb` do not render on GitHub, but the others work fine.

openjournals/joss-reviews#10342
