### 2026-04-07 Machine-Learning-Dynamical-Systems/kooplearn#25 @sdahdah
README should include instructions for installing from source and running unit tests
The project includes great instructions for new users. Instructions to help new contributors set up the development environment would help smooth out the process. Specifically, instructions for installing from source (with jax/torch and development dependencies) and testing would help a lot.

The specific thing that tripped me up was not knowing how to download all the development dependencies in one shot (`pip install ./kooplearn --group dev` was what I ended up doing)

openjournals/joss-reviews#10342
