### 2026-04-07 Machine-Learning-Dynamical-Systems/kooplearn#26 @sdahdah
Minor typos in paper
In the bibliography, proper nouns like Koopman, Hilbert, and Python should be capitalized.

In the title, I would suggest that Scikit-Learn be written in lowercase as scikit-learn (but this is a matter of personal taste I think).

On line 34, "When the dynamics is" should be "When the dynamics are"

Aside from these very minor issues, the text is clear and well-written.

openjournals/joss-reviews#10342
