### 2026-04-07 Machine-Learning-Dynamical-Systems/kooplearn#27 @sdahdah
Three failing unit tests
Three unit tests fail locally for me (using Python 3.14):

```
============================================================================================= short test summary info ==============================================================================================
FAILED tests/test_jax_losses.py::TestVampLoss::test_single_sample - assert Array(False, dtype=bool)
FAILED tests/test_kernel_nystroem_estimator.py::test_callable_kernel_functionality - ValueError: setting an array element with a sequence.
FAILED tests/test_kernel_ridge_estimator.py::test_callable_kernel_functionality - ValueError: setting an array element with a sequence.
```

Are these reproducible on your end?

Full error messages:

```
===================================================================================================== FAILURES =====================================================================================================
_________________________________________________________________________________________ TestVampLoss.test_single_sample __________________________________________________________________________________________

self = <tests.test_jax_losses.TestVampLoss object at 0x7f0556601a50>
single_sample_features = (Array([[-0.02830462,  0.46713185,  0.29570296,  0.15354592, -0.12403282,
         0.21692315, -1.440879  ,  0.7558599....63525754, -1.2226585 ,
        -0.83226097, -0.47417238, -1.2504351 , -0.17678244, -0.04917514]],      dtype=float32))

    def test_single_sample(self, single_sample_features):
        """Test vamp_loss with single sample - JAX returns NaN with centered covariances."""
        x, y = single_sample_features

        # JAX returns NaN with centered covariances for single sample
        loss_centered = vamp_loss(x, y, center_covariances=True)
>       assert jnp.isnan(loss_centered)
E       assert Array(False, dtype=bool)
E        +  where Array(False, dtype=bool) = <PjitFunction of <function isnan at 0x7f0610a3ab90>>(Array(-0., dtype=float32))
E        +    where <PjitFunction of <function isnan at 0x7f0610a3ab90>> = jnp.isnan

tests/test_jax_losses.py:130: AssertionError
________________________________________________________________________________________ test_callable_kernel_functionality ________________________________________________________________________________________
TypeError: only 0-dimensional arrays can be converted to Python scalars

    def test_callable_kernel_functionality():
        X = np.random.randn(15, 5)

        def custom_kernel(X, Y=None):
            # Ensure both X and Y are 2D
            X = np.atleast_2d(X)
            Y = X if Y is None else np.atleast_2d(Y)
            # Compute squared Euclidean distance
            dists = np.sum((X[:, None, :] - Y[None, :, :]) ** 2, axis=-1)
            return np.exp(-0.1 * dists)

        model = NystroemKernelRidge(kernel=custom_kernel, n_components=3)
>       model.fit(X)

tests/test_kernel_nystroem_estimator.py:148:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
src/kooplearn/kernel/_nystroem.py:280: in fit
    kernel_Xnys, kernel_Ynys = self._pre_fit_checks(X)
                               ^^^^^^^^^^^^^^^^^^^^^^^
src/kooplearn/kernel/_nystroem.py:714: in _pre_fit_checks
    self.kernel_X_, self.kernel_Y_, self.kernel_YX_ = self._init_kernels(
src/kooplearn/kernel/_nystroem.py:692: in _init_kernels
    K_X = self._get_kernel(X)
          ^^^^^^^^^^^^^^^^^^^
src/kooplearn/kernel/_nystroem.py:681: in _get_kernel
    return pairwise_kernels(
/home/steven/.virtualenvs/kooplearn/lib/python3.14/site-packages/sklearn/utils/_param_validation.py:218: in wrapper
    return func(*args, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^
/home/steven/.virtualenvs/kooplearn/lib/python3.14/site-packages/sklearn/metrics/pairwise.py:2672: in pairwise_kernels
    return _parallel_pairwise(X, Y, func, n_jobs, **kwds)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
/home/steven/.virtualenvs/kooplearn/lib/python3.14/site-packages/sklearn/metrics/pairwise.py:1969: in _parallel_pairwise
    return func(X, Y, **kwds)
           ^^^^^^^^^^^^^^^^^^
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

X = array([[-0.28232708, -3.91127756, -1.17862478,  0.57909918,  1.22841964],
       [ 0.40255907, -0.24311351,  2.13129145, -0.65239151, -0.91694276]])
Y = array([[-0.28232708, -3.91127756, -1.17862478,  0.57909918,  1.22841964],
       [ 0.40255907, -0.24311351,  2.13129145, -0.65239151, -0.91694276]])
metric = <function test_callable_kernel_functionality.<locals>.custom_kernel at 0x7f04d025a4b0>, ensure_all_finite = True, kwds = {}
xp = <module 'sklearn.externals.array_api_compat.numpy' from '/home/steven/.virtualenvs/kooplearn/lib/python3.14/site-packages/sklearn/externals/array_api_compat/numpy/__init__.py'>
_ = array([[-0.28232708, -3.91127756, -1.17862478,  0.57909918,  1.22841964],
       [ 0.40255907, -0.24311351,  2.13129145, -0.65239151, -0.91694276]]), device = None, dtype_float = <class 'float'>
_get_slice = <function _pairwise_callable.<locals>._get_slice at 0x7f04d025a560>, out = array([[0., 0.],
       [0., 0.]]), iterator = <itertools.combinations object at 0x7f04d020ecf0>

    def _pairwise_callable(X, Y, metric, ensure_all_finite=True, **kwds):
        """Handle the callable case for pairwise_{distances,kernels}."""
        xp, _, device = get_namespace_and_device(X)
        X, Y = check_pairwise_arrays(
            X,
            Y,
            dtype=None,
            ensure_all_finite=ensure_all_finite,
            # No input dimension checking done for custom metrics (left to user)
            ensure_2d=False,
        )
        _, _, dtype_float = _find_floating_dtype_allow_sparse(X, Y, xp=xp)

        def _get_slice(array, index):
            # TODO: below 2 lines can be removed once min scipy >= 1.14. Support for
            # 1D shapes in scipy sparse arrays (COO, DOK and CSR formats) only
            # added in 1.14. We must return 2D array until min scipy 1.14.
            if issparse(array):
                return array[[index], :]
            # When `metric` is a callable, 1D input arrays allowed, in which case
            # scalar should be returned.
            if array.ndim == 1:
                return array[index]
            else:
                return array[index, ...]

        if X is Y:
            # Only calculate metric for upper triangle
            out = xp.zeros((X.shape[0], Y.shape[0]), dtype=dtype_float, device=device)
            iterator = itertools.combinations(range(X.shape[0]), 2)
            for i, j in iterator:
                x = _get_slice(X, i)
                y = _get_slice(Y, j)
>               out[i, j] = metric(x, y, **kwds)
                ^^^^^^^^^
E               ValueError: setting an array element with a sequence.

/home/steven/.virtualenvs/kooplearn/lib/python3.14/site-packages/sklearn/metrics/pairwise.py:2026: ValueError
________________________________________________________________________________________ test_callable_kernel_functionality ________________________________________________________________________________________
TypeError: only 0-dimensional arrays can be converted to Python scalars

    def test_callable_kernel_functionality():
        X = np.random.randn(15, 5)

        def custom_kernel(X, Y=None):
            # Ensure both X and Y are 2D
            X = np.atleast_2d(X)
            Y = X if Y is None else np.atleast_2d(Y)
            # Compute squared Euclidean distance
            dists = np.sum((X[:, None, :] - Y[None, :, :]) ** 2, axis=-1)
            return np.exp(-0.1 * dists)

        model = KernelRidge(kernel=custom_kernel, n_components=3)
>       model.fit(X)

tests/test_kernel_ridge_estimator.py:159:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
src/kooplearn/kernel/_base.py:297: in fit
    self._pre_fit_checks(X)
src/kooplearn/kernel/_base.py:741: in _pre_fit_checks
    self.kernel_X_, self.kernel_Y_, self.kernel_YX_ = self._init_kernels(
src/kooplearn/kernel/_base.py:725: in _init_kernels
    K_X = self._get_kernel(X)
          ^^^^^^^^^^^^^^^^^^^
src/kooplearn/kernel/_base.py:714: in _get_kernel
    return pairwise_kernels(
/home/steven/.virtualenvs/kooplearn/lib/python3.14/site-packages/sklearn/utils/_param_validation.py:218: in wrapper
    return func(*args, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^
/home/steven/.virtualenvs/kooplearn/lib/python3.14/site-packages/sklearn/metrics/pairwise.py:2672: in pairwise_kernels
    return _parallel_pairwise(X, Y, func, n_jobs, **kwds)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
/home/steven/.virtualenvs/kooplearn/lib/python3.14/site-packages/sklearn/metrics/pairwise.py:1969: in _parallel_pairwise
    return func(X, Y, **kwds)
           ^^^^^^^^^^^^^^^^^^
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

X = array([[ 0.05800816,  0.23725996,  0.52901326, -1.43011396,  0.38050519],
       [ 0.93686186,  0.98620571,  1.5514333...813,  0.95487214,  0.73036538,  0.26721906],
       [-0.02347675,  2.08710382, -0.2467946 , -0.65483205,  0.63002319]])
Y = array([[ 0.05800816,  0.23725996,  0.52901326, -1.43011396,  0.38050519],
       [ 0.93686186,  0.98620571,  1.5514333...813,  0.95487214,  0.73036538,  0.26721906],
       [-0.02347675,  2.08710382, -0.2467946 , -0.65483205,  0.63002319]])
metric = <function test_callable_kernel_functionality.<locals>.custom_kernel at 0x7f04cbc68670>, ensure_all_finite = True, kwds = {}
xp = <module 'sklearn.externals.array_api_compat.numpy' from '/home/steven/.virtualenvs/kooplearn/lib/python3.14/site-packages/sklearn/externals/array_api_compat/numpy/__init__.py'>
_ = array([[ 0.05800816,  0.23725996,  0.52901326, -1.43011396,  0.38050519],
       [ 0.93686186,  0.98620571,  1.5514333...813,  0.95487214,  0.73036538,  0.26721906],
       [-0.02347675,  2.08710382, -0.2467946 , -0.65483205,  0.63002319]])
device = None, dtype_float = <class 'float'>, _get_slice = <function _pairwise_callable.<locals>._get_slice at 0x7f04cbc68720>
out = array([[0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0.],
       [0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0... 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0.],
       [0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0.]])
iterator = <itertools.combinations object at 0x7f04cbe3b970>

    def _pairwise_callable(X, Y, metric, ensure_all_finite=True, **kwds):
        """Handle the callable case for pairwise_{distances,kernels}."""
        xp, _, device = get_namespace_and_device(X)
        X, Y = check_pairwise_arrays(
            X,
            Y,
            dtype=None,
            ensure_all_finite=ensure_all_finite,
            # No input dimension checking done for custom metrics (left to user)
            ensure_2d=False,
        )
        _, _, dtype_float = _find_floating_dtype_allow_sparse(X, Y, xp=xp)

        def _get_slice(array, index):
            # TODO: below 2 lines can be removed once min scipy >= 1.14. Support for
            # 1D shapes in scipy sparse arrays (COO, DOK and CSR formats) only
            # added in 1.14. We must return 2D array until min scipy 1.14.
            if issparse(array):
                return array[[index], :]
            # When `metric` is a callable, 1D input arrays allowed, in which case
            # scalar should be returned.
            if array.ndim == 1:
                return array[index]
            else:
                return array[index, ...]

        if X is Y:
            # Only calculate metric for upper triangle
            out = xp.zeros((X.shape[0], Y.shape[0]), dtype=dtype_float, device=device)
            iterator = itertools.combinations(range(X.shape[0]), 2)
            for i, j in iterator:
                x = _get_slice(X, i)
                y = _get_slice(Y, j)
>               out[i, j] = metric(x, y, **kwds)
                ^^^^^^^^^
E               ValueError: setting an array element with a sequence.

/home/steven/.virtualenvs/kooplearn/lib/python3.14/site-packages/sklearn/metrics/pairwise.py:2026: ValueError
```

openjournals/joss-reviews#10342
