### 2026-04-13 Machine-Learning-Dynamical-Systems/kooplearn#28 @gmarupilla
pytest collection fails without optional dependencies (torch/jax)
Opened as part of the JOSS review: openjournals/joss-reviews#10342

## Problem

`pytest.ini` sets:

```ini
addopts = --doctest-modules src
testpaths =
    tests
    src
```

This causes pytest to attempt doctest collection from `src/kooplearn/torch/` and `src/kooplearn/jax/`, which immediately fails with an `ImportError` when the optional extras are not installed:

```
ERROR src/kooplearn/torch/__init__.py
ImportError: To use kooplearn's deep learning losses please reinstall it
with the `torch` extra flag by typing `pip install kooplearn[torch]`.
```

Running `pytest tests/` on a clean install (core only) is therefore broken out of the box. The workaround is `pytest --override-ini="addopts="`.

## Suggested fix

Scope the doctest collection to only the always-available subpackages, for example:

```ini
addopts = --doctest-modules src/kooplearn/linear_model src/kooplearn/kernel
```

Or use a `conftest.py` to conditionally skip collection of optional subpackages when their dependencies are absent.
