### 2026-04-13 Machine-Learning-Dynamical-Systems/kooplearn#29 @gmarupilla
paper.bib: duplicate key and unused references
Opened as part of the JOSS review: openjournals/joss-reviews#10342

## 1. Duplicate BibTeX key `dlkoopman`

`docs/JOSS/paper.bib` contains two entries with the same key `dlkoopman` — one typed `@inproceedings` (line 244) and one `@InProceedings` (line 253). BibTeX/Pandoc will either silently drop one or emit an error during paper compilation. Please merge them into a single entry (keeping whichever is more complete).

## 2. Unused references

The following keys are defined in `paper.bib` but never cited in `paper.md`:

- `kawahara2016`
- `otto2019`
- `Steinwart2008`
- `takeishi2017`

Please either cite them where appropriate or remove them to keep the bibliography clean.
