### 2026-04-13 Machine-Learning-Dynamical-Systems/kooplearn#30 @gmarupilla
Repository & documentation polish
Opened as part of the JOSS review: openjournals/joss-reviews#10342

A collection of small but addressable polish items found during review.

## 1. Typo in contributing filename

The file is named `CONTIBUTING.md` (missing the 'R'). GitHub's UI and most tooling look for `CONTRIBUTING.md` by name. Please rename it.

## 2. Typo in `docs/index.rst`

Line 12: `eigenfucntions` → `eigenfunctions`

## 3. Missing `__version__` attribute

`import kooplearn; kooplearn.__version__` raises `AttributeError`. It is standard Python practice to expose this. Consider adding to `src/kooplearn/__init__.py`:

```python
from importlib.metadata import version
__version__ = version("kooplearn")
```

## 4. No `CODE_OF_CONDUCT.md`

There is no code of conduct file in the repository root. While not a hard JOSS requirement, it is standard open-source practice. The GitHub Contributor Covenant template is a quick addition.

## 5. Thin contributing guide

Once renamed, `CONTRIBUTING.md` only documents the maintainer release process. It does not tell external contributors how to set up a dev environment, run the test suite, or submit a bug report. The README partially covers this — it would be helpful to consolidate that guidance into `CONTRIBUTING.md`.

## 6. Notebooks requiring optional extras lack an install hint

The following notebooks import `torch` or `jax` and fail with an unhelpful error when only the core package is installed:

- `examples/logistic_map.ipynb`
- `examples/ordered_mnist_torch.ipynb`
- `examples/ordered_mnist_jax.ipynb`
- `examples/prinz_potential.ipynb`

A short Markdown cell at the top of each (e.g. `> This notebook requires: pip install "kooplearn[torch]"`) would save users confusion.
