### 2026-04-13 #10342 @gmarupilla
I have completed my review of `kooplearn`. Overall the submission is in good shape — the software is well-designed, thoroughly documented, sklearn-compatible, and makes a genuine contribution to the dynamical systems / machine learning ecosystem.

I verified the following:
- Installation via `pip install kooplearn` works correctly
- Core functionality confirmed: `Ridge`, `KernelRidge`, `NystroemKernelRidge`, `GeneratorDirichlet` all fit/predict/score correctly
- `sklearn.utils.estimator_checks.check_estimator(Ridge())` — all compatibility checks passed
- 6 of 9 example notebooks executed without errors (the remaining 3 require `torch`/`jax` which I could not install in my environment)
- ~2200 tests pass across the ridge/kernel/nystroem/generator/dataset modules

I found two issues I consider blocking, and several smaller polish items. I have filed them as GitHub issues on the repository:

- Machine-Learning-Dynamical-Systems/kooplearn#28 — `pytest.ini` breaks test collection without optional dependencies (**blocking**)
- Machine-Learning-Dynamical-Systems/kooplearn#29 — `paper.bib` duplicate key and unused references (**blocking**)
- Machine-Learning-Dynamical-Systems/kooplearn#30 — Repository & documentation polish (recommended)

I recommend **minor revisions** before acceptance. The blocking issues are straightforward to fix.
