### 2026-04-14 #10342 @jolars
> Hi, I’m a beginner contributor and I’d like to help improve the README by adding installation instructions and steps to run tests. Please let me know if I can work on this.

Thanks @Tushar6405, but this is a review thread for JOSS, so this is not quite the right place. That being said, I am certain that the maintainers of the package would welcome your contributions to the repo at [Machine-Learning-Dynamical-Systems/kooplearn](https://github.com/Machine-Learning-Dynamical-Systems/kooplearn)
