### 2026-04-20 12de68e
docs: fix 'dynamics is'->'dynamics are' and protect Koopman capitalization in bib
- docs/primer.md and docs/JOSS/paper.md: 'When the dynamics is deterministic'
  -> 'When the dynamics are deterministic' (plural 'dynamics' takes 'are').
- docs/JOSS/paper.bib: wrap 'Koopman' in {} in the Williams 2015 title so BibTeX
  styles that downcase titles preserve the proper noun.

Refs #26
