+++ docs/JOSS/paper.bib
@@ -139,7 +139,7 @@
 }
 
 @article{williams2015_kdmd,
-  title = {A kernel-based method for data-driven koopman spectral analysis},
+  title = {A kernel-based method for data-driven {K}oopman spectral analysis},
   journal = {Journal of Computational Dynamics},
   volume = {2},
   number = {2},
