+++ docs/JOSS/paper.md
@@ -68,7 +68,7 @@ bibliography: paper.bib
 
 From fluid flows down to atomistic motions, dynamical systems permeate every scientific discipline. Among the data-driven frameworks for modeling dynamical systems, evolution operator learning [@kostic2022] is both general and principled, and is especially well suited for interpretability [@schutte2001transfer; @mezic2005] and dimensionality reduction [@klus2018data]. An evolution operator $\mathsf{E}$ characterizes dynamical systems, either stochastic $x_{t+1} \sim p(\cdot | x_{t})$, or deterministic $x_{t+1} \sim \delta(\cdot - F(x_{t}))$, as follows: for every function $f$ of the state of the system, $(\mathsf{E} f)(x_{t})$ is the expected value of $f$ one step ahead in the future, given that at time $t$ the system was found in $x_t$
 $$(\mathsf{E} f)(x_t) = \int p(dy | x_{t}) f(y) = \mathbb{E}_{y \sim X_{t+1} | X_{t}}[f(y) | x_t].$$
-Notice that $\mathsf{E}$ is an operator because it maps any function $f$ to another function, $x_{t} \mapsto (\mathsf{E} f)(x_t)$, and is *linear* because $\mathsf{E}(f + \alpha g) = \mathsf{E} f + \alpha \mathsf{E} g$. When the dynamics is deterministic, $\mathsf{E}$ is known as the *Koopman operator* [@koopman1931], while in the stochastic case it is known as the *transfer operator* [@applebaum2009]. Arguably, the most important feature of evolution operators is their spectral decomposition [@mezic2005], which can be used to express the dynamics as a linear superposition of *modes*. These ideas lie at the core of the celebrated Time-lagged Independent Component Analysis [@Molgedey1994], and  Dynamical Mode Decomposition (DMD) [@schmid2010; @kutz2016]. 
+Notice that $\mathsf{E}$ is an operator because it maps any function $f$ to another function, $x_{t} \mapsto (\mathsf{E} f)(x_t)$, and is *linear* because $\mathsf{E}(f + \alpha g) = \mathsf{E} f + \alpha \mathsf{E} g$. When the dynamics are deterministic, $\mathsf{E}$ is known as the *Koopman operator* [@koopman1931], while in the stochastic case it is known as the *transfer operator* [@applebaum2009]. Arguably, the most important feature of evolution operators is their spectral decomposition [@mezic2005], which can be used to express the dynamics as a linear superposition of *modes*. These ideas lie at the core of the celebrated Time-lagged Independent Component Analysis [@Molgedey1994], and  Dynamical Mode Decomposition (DMD) [@schmid2010; @kutz2016]. 
 
 Evolution operator learning is best understood from the perspective of *latent linear dynamical models*, which is schematically depicted in \autoref{fig:evop_scheme}. In this framework, the dynamical state $x_t$ is first mapped into a latent space defined by a (fixed or learned) representation $\varphi$. Then, a *linear evolution* map $E$ is learned to approximate the dynamics of the latents. The pair $(\varphi, E)$ provides an approximation of $\mathsf{E}$ restricted to the $d$-dimensional subspace spanned by the components of $\varphi$, given the data. `kooplearn` implements methods to learn $\varphi$, $E$, and the associated spectral decomposition of $\mathsf{E}$. 
 
