### 2026-04-25 #10342 @sepandhaghighi
@g-turri Firstly, I would like to express my gratitude for submitting this work to the **JOSS** journal.

After conducting a review of `kooplearn`, I conclude that it is a well-designed project with a significant impact on the dynamical systems ecosystem. However, I identified two issues that are blockers and some suggestions for future development.

### Issues

1. The current contribution guidelines are insufficient for a project of this scope. I recommend creating a more comprehensive guideline that outlines the different modules, development disciplines, and procedures for adding new features. This information should be documented in the `CONTRIBUTING.md` file, while the maintainer's release notes can be moved to a separate file, such as `RELEASE.md`.

2. Due to the project's various dependencies, it would be beneficial to expand the installation section to cover different installation modes, including steps for installing from source, and running tests.

### Suggestions

1. Reporting project test coverage statistics, which is crucial for assessing the project's test quality.

2. Adding a collection of issue templates for reporting bugs, new features, and other issues.

3. Adding a pull request template for streamlined submission and review processes.

----------
* I used [MyText v0.7](https://github.com/sepandhaghighi/mytext) (model: `meta/llama-3-8b-instruct`) to check the grammar in my review text.
