### 2026-05-06 7b9a1a4
fix(tests): avoid optional dependency collection failures
