+++ pytest.ini
@@ -1,6 +1,7 @@
 # pytest.ini
 [pytest]
-addopts = --doctest-modules src
-testpaths = 
+addopts = --doctest-modules
+testpaths =
     tests
-    src
+    src/kooplearn/linear_model
+    src/kooplearn/kernel
