### 2026-05-06 076cbd6
fix(tests): stabilize reviewer-reported edge cases
