### 2026-05-06 44a3fe6
fix(paper): clean JOSS bibliography and typos
