+++ docs/JOSS/paper.bib
@@ -42,7 +42,7 @@
 
 @article{sklearn,
   author = {Pedregosa, Fabian and Varoquaux, Ga\"{e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and Vanderplas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, Matthieu and Duchesnay, \'{E}douard},
-  title = {Scikit-learn: Machine Learning in Python},
+  title = {{Scikit-learn}: Machine Learning in {Python}},
   year = {2011},
   issue_date = {2/1/2011},
   publisher = {JMLR.org},
@@ -81,22 +81,6 @@
   pages = {3634–3637}
 }
 
-@inbook{Steinwart2008,
-  author = {Ingo Steinwart and Andreas Christmann},
-  title = {Support Vector Machines},
-  booktitle = {Least Squares Support Vector Machines},
-  chapter = {},
-  pages = {29-70},
-  doi = {10.1142/9789812776655_0002},
-  URL = {https://www.worldscientific.com/doi/abs/10.1142/9789812776655_0002},
-  eprint = {https://www.worldscientific.com/doi/pdf/10.1142/9789812776655_0002},
-  abstract = { Abstract The following sections are included: Maximal margin classification and linear SVMs Margin Linear SVM classifier: separable case Linear SVM classifier: non-separable case Kernel trick and Mercer condition Nonlinear SVM classifiers VC theory and structural risk minimization Empirical risk versus generalization error Structural risk minimization SVMs for function estimation SVM for linear function estimation SVM for nonlinear function estimation VC bound on generalization error Modifications and extensions Kernels Extension to other convex cost functions Algorithms Parametric versus non-parametric approaches },
-  city = {New York, NY},
-  publisher = {Springer New York},
-  title = {Support Vector Machines},
-  year = {2008},
-}
-
 @article{klus2018data,
   author  = {Klus, Stefan and N{\"u}ske, Feliks and Koltai, P{\'e}ter and Wu, Hao and Kevrekidis, Ioannis and Sch{\"u}tte, Christof and No{\'e}, Frank},
   title   = {Data-Driven Model Reduction and Transfer Operator Approximation},
@@ -123,7 +107,9 @@
   address="Berlin, Heidelberg",
   pages="191--223",
   abstract="The article surveys the development of novel mathematical concepts and algorithmic approaches based thereon in view of their possible applicability to biomolecular design. Both a first deterministic approach, based on the Frobenius-Perron operator corresponding to the flow of the Hamiltonian dynamics, and later stochastic approaches, based on a spatial Markov operator or on Langevin dynamics, can be subsumed under the unified mathematical roof of the transfer operator approach to effective dynamics of molecular systems. The key idea of constructing specific transfer operators especially taylored for the purpose of conformational dynamics appears as the red line throughout the paper. Different steps of the algorithm are exemplified by a trinucleotide molecular system as a small representative of possible RNA drug molecules.",
-  isbn="978-3-642-56589-2"
+  isbn="978-3-642-56589-2",
+  doi="10.1007/978-3-642-56589-2_9",
+  url="https://doi.org/10.1007/978-3-642-56589-2_9"
 }
 
 @book{kutz2016,
@@ -139,7 +125,7 @@
 }
 
 @article{williams2015_kdmd,
[…]
