### 2026-05-06 a2a95ad
chore(repo): polish contributor docs and project metadata
