+++ .github/pull_request_template.md
@@ -1,0 +1,18 @@
+## Summary
+
+Describe the change and why it is needed.
+
+## Verification
+
+List the commands you ran, for example:
+
+```bash
+uv run pytest
+ruff check
+```
+
+## Checklist
+
+- [ ] Tests or documentation were updated when behavior changed.
+- [ ] Optional dependency changes are documented.
+- [ ] The change is focused and ready for review.
