+++ README.md
@@ -64,19 +64,40 @@ uv add "kooplearn[torch]"
 uv add "kooplearn[jax]"
 ```
 
+### From source
+
+For development, clone the repository and install the package with all optional extras and dependency groups:
+
+```bash
+git clone https://github.com/Machine-Learning-Dynamical-Systems/kooplearn.git
+cd kooplearn
+uv sync --all-extras --all-groups
+```
+
+With `pip>=25.1`, the equivalent editable install is:
+
+```bash
+python -m pip install -U pip
+python -m pip install -e ".[torch,jax]" --group dev --group docs --group examples
+```
+
+## Testing
+
+Run the default test suite from the repository root with:
+
+```bash
+uv run pytest
+```
+
+After installing with `pip`, use:
+
+```bash
+python -m pytest
+```
+
 ## Contributing
 
-We welcome contributions from the community! If you're interested in contributing to `kooplearn`, please follow these steps:
-
-1. Fork the repository on GitHub.
-2. Clone your forked repository to your local machine.
-3. Create a new branch for your feature or bug fix: `git checkout -b feature/your-feature-name` or `git checkout -b bugfix/issue-number`.
-4. Make your changes and commit them with descriptive commit messages.
-5. Push your changes to your forked repository.
-6. Create a pull request from your branch to the `main` branch of the original repository.
-7. Provide a clear title and description for your pull request, including any relevant information about the changes you've made.
-
-We appreciate your contributions and will review your pull request as soon as possible. Thank you for helping improve `kooplearn`!
+We welcome contributions from the community. See [CONTRIBUTING.md](CONTRIBUTING.md) for development setup, testing, issue reports, and pull request guidance.
 
 ## License
 
@@ -101,7 +122,7 @@ The mantainer of this repo is Pietro Novelli.
 
 ```bibtex
 @article{kooplearn,
-title={kooplearn: A Scikit-Learn Compatible Library of Algorithms for Evolution Operator Learning}, 
+title={kooplearn: A scikit-learn compatible library of algorithms for evolution operator learning},
 author={Giacomo Turri and Grégoire Pacreau and Giacomo Meanti and Timothée Devergne and Daniel Ordonez and Erfan Mirzaei and Bruno Belucci and Karim Lounici and Vladimir R. Kostic and Massimiliano Pontil and Pietro Novelli},
 year={2026},
 eprint={2512.21409},
@@ -114,4 +135,3 @@ url={https://arxiv.org/abs/2512.21409},
 ---
 
 We hope you find `kooplearn` useful for your dynamical systems analysis. If you encounter any issues or have suggestions for improvements, please don't hesitate to [raise an issue](https://github.com/Machine-Learning-Dynamical-Systems/kooplearn/issues). Happy coding!
-
