+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,20 @@
+# Code of Conduct
+
+We are committed to a welcoming, respectful, and productive community around `kooplearn`.
+
+## Expected behavior
+
+- Be respectful and constructive in issues, pull requests, reviews, and discussions.
+- Focus criticism on ideas and code, not on people.
+- Assume good intent, while accepting responsibility for the impact of your words and actions.
+- Respect different backgrounds, experience levels, and viewpoints.
+
+## Unacceptable behavior
+
+Harassment, discriminatory language, personal attacks, sustained disruption, and publishing private information without permission are not acceptable.
+
+## Reporting
+
+Report conduct concerns privately to the maintainers listed in the repository metadata or through the contact channel used by the project. Maintainers will review reports, protect reporter privacy as much as practical, and take proportionate action when needed.
+
+This policy follows the spirit of the [Contributor Covenant](https://www.contributor-covenant.org/).
