### 2026-05-06 4aaf5d8
fix(paper): use sentence case title
