### 2026-05-06 fbb5f6b
chore(notebooks): clean GitHub rendering metadata
