### 2026-05-06 4d61b32
ci(coverage): add codecov reporting
