+++ pyproject.toml
@@ -29,6 +29,7 @@ build-backend = "hatchling.build"
 dev = [
     "ipykernel>=6.30.1",
     "pytest>=8.4.2",
+    "pytest-cov>=7.1.0",
     "ruff>=0.13.2",
     "ty>=0.0.1a21",
 ]
