+++ .github/workflows/CI.yml
@@ -45,3 +45,31 @@ jobs:
       - name: Run tests
         # For example, using `pytest`
         run: uv run pytest --color=yes tests/
+
+  coverage:
+    name: Coverage on ubuntu-latest, Python 3.12
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+      id-token: write
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@v6
+        with:
+          python-version: "3.12"
+
+      - name: Install the project
+        run: uv sync --locked --all-extras --all-groups
+
+      - name: Run coverage
+        run: uv run pytest --color=yes --cov=kooplearn --cov-report=xml tests/
+
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v5
+        with:
+          fail_ci_if_error: true
+          files: ./coverage.xml
+          use_oidc: true
