### 2026-05-06 6f76b87
docs(readme): add project badges
