### 2026-05-06 4b3a7e5
docs(readme): move logo assets into docs static
