### 2026-05-06 Machine-Learning-Dynamical-Systems/kooplearn#24 @g-turri
Hi @sdahdah, thanks for spotting this!

I addressed this in PR #32, specifically in commit Machine-Learning-Dynamical-Systems/kooplearn@fbb5f6b. The notebooks `ala2_nys_tutorial.ipynb` and `kernel_methods.ipynb` should now render correctly on GitHub.

Please let me know if this resolves the issue on your side.
