### 2026-05-06 Machine-Learning-Dynamical-Systems/kooplearn#25 @g-turri
Thanks for raising this and for pointing out that the source-install and testing instructions were not clear enough — I addressed this in Machine-Learning-Dynamical-Systems/kooplearn@a2a95ad.

The README now includes source-installation instructions for both `uv` and editable `pip` installs, including optional extras and dependency groups. It also documents the test commands using both `uv run pytest` and `python -m pytest`.

I also expanded `CONTRIBUTING.md`, so contributor setup and testing guidance is now available directly in the repository documentation.
