### 2026-05-06 Machine-Learning-Dynamical-Systems/kooplearn#26 @g-turri
Thanks for catching these formatting and copy-editing issues.

I addressed them in Machine-Learning-Dynamical-Systems/kooplearn@44a3fe6 and Machine-Learning-Dynamical-Systems/kooplearn@4aaf5d8. I fixed the grammar issue, corrected bibliography capitalization, and updated the paper title to sentence case.
