### 2026-05-06 Machine-Learning-Dynamical-Systems/kooplearn#28 @g-turri
Thanks @gmarupilla for pointing this out.

I addressed this in Machine-Learning-Dynamical-Systems/kooplearn@7b9a1a4. I scoped doctest collection in `pytest.ini` to the always-available core subpackages and added `pytest.importorskip(...)` guards to the optional torch/JAX test modules. This prevents collection from failing when optional extras are not installed.
