### 2026-05-06 Machine-Learning-Dynamical-Systems/kooplearn#29 @g-turri
Thanks for catching the bibliography issues.

I addressed them in Machine-Learning-Dynamical-Systems/kooplearn@44a3fe6. I removed the duplicate `dlkoopman` entry, removed unused bibliography entries, and filled in the reviewer/editorialbot DOI gaps where I could verify them. I also ran a citation audit afterward to confirm there are no duplicate keys, missing citations, or unused references left.
