### 2026-05-06 Machine-Learning-Dynamical-Systems/kooplearn#30 @g-turri
Thanks for the detailed suggestions on improving the repository structure and documentation for contributors.

I addressed them in Machine-Learning-Dynamical-Systems/kooplearn@a2a95ad and Machine-Learning-Dynamical-Systems/kooplearn@c9e1ce0. I renamed `CONTIBUTING.md` to `CONTRIBUTING.md`, expanded the contributor setup/testing/reporting guidance, and moved the maintainer release process into `RELEASE.md`.

I also added `CODE_OF_CONDUCT.md`, issue templates, a PR template, `kooplearn.__version__`, the `docs/index.rst` typo fix, and install-hint cells for the optional-dependency notebooks.
