### 2026-05-06 #10342 @g-turri
> My overall comment is that the project is well-structured and carefully designed. It is a significant contribution to the field. Once the minor issues I've raised have been addressed I'd be happy to recommend acceptance.

Hi @sdahdah, thank you very much for the careful review and for the positive assessment of the project.

I have addressed the issues you raised in Machine-Learning-Dynamical-Systems/kooplearn#32. In particular:

- Machine-Learning-Dynamical-Systems/kooplearn@fbb5f6b cleans the notebooks `ala2_nys_tutorial.ipynb` and `kernel_methods.ipynb`, which should now render correctly on GitHub.

- Machine-Learning-Dynamical-Systems/kooplearn@a2a95ad expands the README and `CONTRIBUTING.md` with clearer source installation and testing instructions.

- Machine-Learning-Dynamical-Systems/kooplearn@44a3fe6 and Machine-Learning-Dynamical-Systems/kooplearn@4aaf5d8 address the paper typo, title casing, and bibliography capitalization issues.

- Machine-Learning-Dynamical-Systems/kooplearn@076cbd6 fixes the reported test failures related to callable-kernel behavior and the single-sample centered JAX-loss assertion.

I also replied directly to the corresponding issues with more details about each change. Please let me know if these fixes resolve the issues on your side.
