### 2026-05-07 Machine-Learning-Dynamical-Systems/kooplearn#29 @gmarupilla
Thanks for addressing it.
