### 2026-05-07 Machine-Learning-Dynamical-Systems/kooplearn#26 @sdahdah
Thanks for addressing
