### 2026-05-07 Machine-Learning-Dynamical-Systems/kooplearn#25 @sdahdah
Looks great, thank you
