### 2026-05-07 Machine-Learning-Dynamical-Systems/kooplearn#27 @sdahdah
I was able to run the tests, but `loguru` and `tyro` are missing from the `pyproject.toml` dev dependencies, so I had to install them manually. Could you please add them?
