### 2026-05-09 #10342 @sepandhaghighi
> Hi [@sepandhaghighi](https://github.com/sepandhaghighi), thank you for the thoughtful review and for the concrete suggestions on improving the project’s contributor workflow, installation guidance, and review infrastructure.
>
> Since your feedback was posted directly in this review thread rather than as separate repository issues, I am summarizing the corresponding changes here.

@g-turri, well done! 🥇
I took a look at the PR, and everything looks good to me 🔥

@jolars I am satisfied with the current state of the project, and my suggestion is to **accept** this submission 💯
