### 2026-05-11 0f78e27
fix(pyproject): declare missing runtime and dev dependencies
