+++ pyproject.toml
@@ -8,7 +8,9 @@ authors = [
 ]
 requires-python = ">=3.10"
 dependencies = [
+    "numpy>=2.0.2",
     "pandas>=2.3.3",
+    "scipy>=1.13.1",
     "scikit-learn>=1.6.1",
 ]
 
@@ -28,10 +30,12 @@ build-backend = "hatchling.build"
 [dependency-groups]
 dev = [
     "ipykernel>=6.30.1",
+    "loguru>=0.7.3",
     "pytest>=8.4.2",
     "pytest-cov>=7.1.0",
     "ruff>=0.13.2",
     "ty>=0.0.1a21",
+    "tyro>=0.9.35",
 ]
 docs = [
     "myst-nb>=1.3.0",
