### 2026-05-11 Machine-Learning-Dynamical-Systems/kooplearn#27 @g-turri
Thanks for catching this! I added `loguru` and `tyro` to the dev dependencies in `pyproject.toml` and updated `uv.lock` accordingly (see Machine-Learning-Dynamical-Systems/kooplearn@0f78e27). I also verified that a fresh dev environment installs successfully.
