### 2026-05-11 Machine-Learning-Dynamical-Systems/kooplearn#27 @sdahdah
Thank you!
