### 2026-05-19 fa62981
docs: trim double-space
