### 2026-05-19 4615340
docs: consistently use "are" for "dynamics"
