### 2026-05-19 a313875
docs: avoid ambiguous "which is"
It sits close to the plural, and reads awkwardly as a result.
