### 2026-05-19 a5e6656
docs: lowercase "Kernel" to "kernel"
