### 2026-05-19 05fe7dc
docs: use bare citation
