+++ docs/JOSS/paper.md
@@ -96,7 +96,7 @@ In continuous-time dynamics, the system's evolution operator can be expressed as
 
 ## Datasets
 
-To foster reproducibility and rigorous benchmarking, `kooplearn` includes the `kooplearn.datasets` module, containing utilities to easily generate trajectories for systems that range from deterministic chaos (e.g., *Lorenz-63*, *Duffing oscillator*, *Logistic Map*) to stochastic and metastable dynamics (e.g., *stochastic linear systems*, *regime-switching models*, *Langevin dynamics*). A distinguishing feature of the library is the inclusion of benchmarks with accessible ground-truth spectral decompositions—such as the *Noisy Logistic Map* [@ostruszka2000dynamical] and *Overdamped Langevin Dynamics* in a quadruple-well potential [@Prinz2011]. These allow users to quantify the accuracy of learned eigenvalues and eigenfunctions directly (as demonstrated in \autoref{fig:eigfns_approximation}). Finally, the suite includes the *Ordered MNIST* from [@kostic2022] to evaluate performance on high-dimensional structured data. Examples of trajectories generated using the `kooplearn.datasets` module are illustrated in \autoref{fig:datasets}.
+To foster reproducibility and rigorous benchmarking, `kooplearn` includes the `kooplearn.datasets` module, containing utilities to easily generate trajectories for systems that range from deterministic chaos (e.g., *Lorenz-63*, *Duffing oscillator*, *Logistic Map*) to stochastic and metastable dynamics (e.g., *stochastic linear systems*, *regime-switching models*, *Langevin dynamics*). A distinguishing feature of the library is the inclusion of benchmarks with accessible ground-truth spectral decompositions—such as the *Noisy Logistic Map* [@ostruszka2000dynamical] and *Overdamped Langevin Dynamics* in a quadruple-well potential [@Prinz2011]. These allow users to quantify the accuracy of learned eigenvalues and eigenfunctions directly (as demonstrated in \autoref{fig:eigfns_approximation}). Finally, the suite includes the *Ordered MNIST* from @kostic2022 to evaluate performance on high-dimensional structured data. Examples of trajectories generated using the `kooplearn.datasets` module are illustrated in \autoref{fig:datasets}.
 
 ![Samples from the datasets included in `kooplearn`.\label{fig:datasets}](Fig4.png){ width=100% }
 
