### 2026-05-19 f5e2c31
docs: use Pandoc-native chapter reference
