+++ docs/JOSS/paper.md
@@ -92,7 +92,7 @@ The ecosystem of Python libraries that support operator-based modeling has grown
 
 ## Learning the Infinitesimal Generator of Diffusion Processes
 
-In continuous-time dynamics, the system's evolution operator can be expressed as the exponential of the *infinitesimal generator* $\mathsf{L}$, a differential operator defined by the equations of motion [@applebaum2009] (Chapter 3). Formally, for time-homogeneous dynamics, the generator relates to the evolution operator via $\mathsf{E} = e^{\mathsf{L}}$, and consequently $\mathbb{E}[f(X_t)\vert x_0] = (e^{t \mathsf{L}}f)(x_0)$. Since the exponential of an operator preserves its eigenfunctions, one can use knowledge of $\mathsf{L}$ (or its properties) to learn dynamical behavior without requiring lag-time data. In other words, it becomes possible to construct a physics-informed kinetic model $\mathsf{E}$ solely from static (equilibrium) data. To this end, `kooplearn` provides implementations of recent kernel-based algorithms for diffusion processes with Dirichlet boundary conditions from [@kostic2024generator], as well as neural representations as proposed in [@devergne2024biased]. As demonstrated in [@devergne2025slow], these approaches improve sample complexity compared to estimators that rely solely on lag-time trajectory data.
+In continuous-time dynamics, the system's evolution operator can be expressed as the exponential of the *infinitesimal generator* $\mathsf{L}$, a differential operator defined by the equations of motion [@applebaum2009, Chapter 3]. Formally, for time-homogeneous dynamics, the generator relates to the evolution operator via $\mathsf{E} = e^{\mathsf{L}}$, and consequently $\mathbb{E}[f(X_t)\vert x_0] = (e^{t \mathsf{L}}f)(x_0)$. Since the exponential of an operator preserves its eigenfunctions, one can use knowledge of $\mathsf{L}$ (or its properties) to learn dynamical behavior without requiring lag-time data. In other words, it becomes possible to construct a physics-informed kinetic model $\mathsf{E}$ solely from static (equilibrium) data. To this end, `kooplearn` provides implementations of recent kernel-based algorithms for diffusion processes with Dirichlet boundary conditions from [@kostic2024generator], as well as neural representations as proposed in [@devergne2024biased]. As demonstrated in [@devergne2025slow], these approaches improve sample complexity compared to estimators that rely solely on lag-time trajectory data.
 
 ## Datasets
 
