### 2026-05-19 18ca6ac
docs: use bare citations, change prepositions
