### 2026-05-19 Machine-Learning-Dynamical-Systems/kooplearn#34 @jolars
Harmonize paper title convention in bibliography
Right now you're mixing sentence and title case for the title in the bibliography. Please pick one (JOSS does not have a preference for this, so feel free to choose which one you prefer.)

Ref: openjournals/joss-reviews#10342
