### 2026-05-19 Machine-Learning-Dynamical-Systems/kooplearn#35 @jolars
Use vector images in the JOSS paper
The images in your paper are all PNG files, but these have lower level of detail, work poorly with screen readers, and take up more space. It may not be possible for all the images, but please convert at least some of them into vector images (PDFs).

Ref: openjournals/joss-reviews#10342
