### 2026-05-19 14cc6f7
docs: initialize name in bibliography
