+++ docs/JOSS/paper.bib
@@ -359,7 +359,7 @@
 }
 
 @inproceedings{kostic2024learning,
-  author = {Kostic, Vladimir R. and Novelli, Pietro and Grazzi, Riccardo and Lounici, Karim and pontil, massimiliano},
+  author = {Kostic, Vladimir R. and Novelli, Pietro and Grazzi, Riccardo and Lounici, Karim and Pontil, Massimiliano},
   booktitle = {International Conference on Representation Learning},
   editor = {B. Kim and Y. Yue and S. Chaudhuri and K. Fragkiadaki and M. Khan and Y. Sun},
   pages = {54329--54341},
