+++ docs/JOSS/paper.md
@@ -106,7 +106,7 @@ To foster reproducibility and rigorous benchmarking, `kooplearn` includes a `koo
 
 The library's architecture is modular, with subpackages for different modeling approaches: `kooplearn.linear_model`, `kooplearn.kernel`, and deep learning utilities in `kooplearn.torch` and `kooplearn.jax`. The estimators within the `linear_model` and `kernel` subpackages, such as `Ridge` and `KernelRidge`, inherit from `sklearn.base.BaseEstimator`, strictly adhering to the `scikit-learn` API. This design choice ensures that users familiar with `scikit-learn` can easily adopt `kooplearn` into their workflows, using methods like `fit`, `predict`, and `score`. To ensure high performance in learning the estimators, the library leverages optimized `scipy.linalg` and `scipy.sparse.linalg` [@scipy] routines, wrapping standard libraries like LAPACK and ARPACK, and offers a flexible selection of solvers. Users can choose from `dense`, `arpack`, or `randomized` strategies to optimize computational efficiency according to the problem size.
 
-For deep learning applications, `kooplearn` provides a suite of theoretically-grounded loss functions, such as the `VampLoss` and `SpectralContrastiveLoss`, which are implemented as `torch.nn.Module` objects. This allows seamless integration with custom neural network architectures in `PyTorch` or `JAX`, enabling users to leverage GPUs for accelerated training.  This combination of a high-level, user-friendly API with low-level computational flexibility makes `kooplearn` a powerful tool for both rapid prototyping and advanced research.
+For deep learning applications, `kooplearn` provides a suite of theoretically-grounded loss functions, such as the `VampLoss` and `SpectralContrastiveLoss`, which are implemented as `torch.nn.Module` objects. This allows seamless integration with custom neural network architectures in `PyTorch` or `JAX`, enabling users to leverage GPUs for accelerated training. This combination of a high-level, user-friendly API with low-level computational flexibility makes `kooplearn` a powerful tool for both rapid prototyping and advanced research.
 
 # Research Impact Statement
 
