### 2026-05-19 866376e
docs: use SI unit convention
