### 2026-05-19 1963d7c
docs: drop tautological "memory"
