### 2026-05-19 96e139f
docs: drop math mode for "$5000$"
