### 2026-05-19 7ca8e73
docs: decapitalize "Transform"
It's used in lowercase elsewhere, so this keeps it consistent.
