+++ docs/JOSS/paper.md
@@ -62,7 +62,7 @@ bibliography: paper.bib
 
 # Summary
 
-`kooplearn` is a machine learning library that implements linear, kernel, and deep learning estimators of *dynamical operators* and their spectral decompositions. `kooplearn` can model both discrete-time evolution operators (Koopman/Transfer) and continuous-time infinitesimal generators. By learning these operators, users can analyze dynamical systems via spectral methods, derive data-driven reduced-order models, and forecast future states and observables. `kooplearn`'s interface is compliant with the `scikit-learn` API [@sklearn], facilitating its integration into existing machine learning and data science workflows. Additionally, `kooplearn` includes curated benchmark datasets to support experimentation, reproducibility, and the fair comparison of learning algorithms. The software is available at <https://github.com/Machine-Learning-Dynamical-Systems/kooplearn>.
+`kooplearn` is a machine learning library that implements linear, kernel, and deep learning estimators of *dynamical operators* and their spectral decompositions. `kooplearn` can model both discrete-time evolution operators (Koopman/transfer) and continuous-time infinitesimal generators. By learning these operators, users can analyze dynamical systems via spectral methods, derive data-driven reduced-order models, and forecast future states and observables. `kooplearn`'s interface is compliant with the `scikit-learn` API [@sklearn], facilitating its integration into existing machine learning and data science workflows. Additionally, `kooplearn` includes curated benchmark datasets to support experimentation, reproducibility, and the fair comparison of learning algorithms. The software is available at <https://github.com/Machine-Learning-Dynamical-Systems/kooplearn>.
 
 # Statement of Need
 
