### 2026-05-27 45f0f43
docs: fix and capitalize "Dynamic Mode Decomposition" in multiple files
