+++ docs/JOSS/paper.bib
@@ -135,7 +135,7 @@
   doi = {10.3934/jcd.2015005},
   url = {https://www.aimsciences.org/article/id/ce535396-f8fe-4aa1-b6de-baaf986f6193},
   author = {Matthew O.  Williams and Clarence W. Rowley and Ioannis G.  Kevrekidis},
-  keywords = {Koopman operator, dynamic mode decomposition, kernel
+  keywords = {Koopman operator, Dynamic Mode Decomposition, kernel
     methods, time series analysis, machine learning}
 }
 
