+++ docs/bibliography.bib
@@ -139,7 +139,7 @@
 }
 
 @article{li2017extended,
-  title={Extended dynamic mode decomposition with dictionary learning: A data-driven adaptive spectral decomposition of the Koopman operator},
+  title={Extended Dynamic Mode Decomposition with dictionary learning: A data-driven adaptive spectral decomposition of the Koopman operator},
   author={Li, Qianxiao and Dietrich, Felix and Bollt, Erik M and Kevrekidis, Ioannis G},
   journal={Chaos: An Interdisciplinary Journal of Nonlinear Science},
   volume={27},
@@ -295,7 +295,7 @@ series = {NIPS'18}
 }
 
 @article{korda2018convergence,
-  title={On convergence of extended dynamic mode decomposition to the {K}oopman operator},
+  title={On convergence of extended Dynamic Mode Decomposition to the {K}oopman operator},
   author={Korda, Milan and Mezi{\'c}, Igor},
   journal={Journal of Nonlinear Science},
   volume={28},
@@ -325,7 +325,7 @@ series = {NIPS'18}
   doi = {10.3934/jcd.2015005},
   url = {https://www.aimsciences.org/article/id/ce535396-f8fe-4aa1-b6de-baaf986f6193},
   author = {Matthew O.  Williams and Clarence W. Rowley and Ioannis G.  Kevrekidis},
-  keywords = {Koopman operator, dynamic mode decomposition, kernel
+  keywords = {Koopman operator, Dynamic Mode Decomposition, kernel
     methods, time series analysis, machine learning}
 }
 
