### 2026-05-27 Machine-Learning-Dynamical-Systems/kooplearn#34 @g-turri
Thanks, that makes sense. I’ve standardized the bibliography titles to sentence case in https://github.com/Machine-Learning-Dynamical-Systems/kooplearn/commit/fa730168d975bb4c648f7a73d4e661860ab2d5e3.
