### 2026-05-27 Machine-Learning-Dynamical-Systems/kooplearn#35 @g-turri
I replaced the PNG figures with PDF versions in https://github.com/Machine-Learning-Dynamical-Systems/kooplearn/commit/b3636b19cc9704e5dac8b5beb40c45b57aeceefa.
