### 2026-05-27 #10342 @g-turri
Hi @jolars, thanks for the edits! I've now merged Machine-Learning-Dynamical-Systems/kooplearn#36 and addressed the issues you raised.
